Part 1 of the pumpkine fix
Added a new layer "bounds" that only reacts to the player's clip
manager's raycasts. this can be used to set bounds that do not interact
with the sensors of the player. said bounds have to be used very
carefully.
Added fix to the movement modifier to display correct animation after
player has collided with pumpkine.